add loand and repayment policy update views

+++ app/Components/CoreComponent/Modules/Loan/LoanController.php
@@ -14,24 +14,36 @@ use Illuminate\Support\Facades\View;
  */
 class LoanController extends Controller
 {
+    protected function getRequiredData()
+    {
+        return [
+            'clientModelClass' => Client::class,
+            'loanModelClass' => Loan::class,
+            'repaymentModelClass' => Repayment::class,
+        ];
+    }
+
     public function index(Request $request)
     {
+        $this->authorize('view', Loan::class);
         $loans = Loan::paginate();
         return $this->view('index', [
             'loans' => $loans,
-        ]);
+        ], $this->getRequiredData());
     }
     public function create(Request $request)
     {
+        $this->authorize('create', Loan::class);
         $client = Client::find($request->get('client_id'));
         $freqTypes = RepaymentFrequency::getFreqType();
         return $this->view('create', [
             'freqTypes' => $freqTypes,
             'client' => $client,
-        ]);
+        ], $this->getRequiredData());
     }
     public function store(Request $request)
     {
+        $this->authorize('create', Loan::class);
         $loan = new Loan();
         $loan->fill($request->except('_token'));
         if ($loan->save()) {
@@ -42,22 +54,25 @@ class LoanController extends Controller
     public function show(Request $request, $id)
     {
         $loan = Loan::find($id);
+        $this->authorize('view', Loan::class);
         return $this->view('show', [
             'loan' => $loan,
-        ]);
+        ], $this->getRequiredData());
     }
     public function edit(Request $request, $id)
     {
         $freqTypes = RepaymentFrequency::getFreqType();
         $loan = Loan::find($id);
+        $this->authorize('update', $loan);
         return $this->view('edit', [
             'freqTypes' => $freqTypes,
             'loan' => $loan,
-        ]);
+        ], $this->getRequiredData());
     }
     public function update(Request $request, $id)
     {
         $loan = Loan::find($id);
+        $this->authorize('update', $loan);
         $loan->fill($request->all());
         if ($loan->save()) {
             return redirect()->route('clients.show', $loan->client_id)->withSuccess('loan have been updated');
@@ -67,6 +82,7 @@ class LoanController extends Controller
     public function destroy(Request $request, $id)
     {
         $loan = Loan::find($id);
+        $this->authorize('delete', $loan);
         if ($loan->delete()) {
             return redirect()->route('loans.index')->withSuccess('loan have been deleted');
         }
@@ -75,7 +91,9 @@ class LoanController extends Controller
 
     public function pay(Request $request, $id)
     {
-        if (Repayment::pay($id)) {
+        $repayment = Repayment::find($id);
+        $this->authorize('pay', $repayment);
+        if ($repayment->pay()) {
             return back()->withSuccess("repayment id $id has been paid");
         }
         return back()->withError("paid fail for repayment id $id");
